Exception validation for Mailbox Profile.

+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/ProfileConfigurationService.java
@@ -25,6 +25,7 @@ import com.liaison.mailbox.service.dto.ResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.request.AddProfileToMailBoxRequestDTO;
 import com.liaison.mailbox.service.dto.configuration.response.AddProfileToMailBoxResponseDTO;
 import com.liaison.mailbox.service.dto.configuration.response.DeactivateMailboxProfileLinkResponseDTO;
+import com.liaison.mailbox.service.exception.MailBoxServicesException;
 
 /**
  * Class which has configuration related operations.
@@ -43,37 +44,53 @@ public class ProfileConfigurationService {
 	 *            The request DTO & guid.
 	 * @return The responseDTO.
 	 */
-	public AddProfileToMailBoxResponseDTO addProfileToMailBox(AddProfileToMailBoxRequestDTO request, String guid) {
+	public AddProfileToMailBoxResponseDTO addProfileToMailBox(AddProfileToMailBoxRequestDTO request, String mailboxGuid) {
 
 		LOG.info("Add Profile to mailbox");
-
-		ScheduleProfilesRef profile = new ScheduleProfilesRef();
-
-		// Getting the existing mailbox from given GUID
-		MailBoxConfigurationDAO config = new MailBoxConfigurationDAOBase();
-		MailBox mailbox = config.find(MailBox.class, guid);
-
-		// Construct MailBoxSchedProfile entity
-		MailBoxSchedProfile mailBoxSchedProfile = new MailBoxSchedProfile();
-		mailBoxSchedProfile.setScheduleProfilesRef(profile);
-		mailBoxSchedProfile.setMailbox(mailbox);
-		request.copyToEntity(mailBoxSchedProfile);
-
-		// Persist the entity
-		MailBoxScheduleProfileConfigurationDAO configDao = new MailBoxScheduleProfileConfigurationDAOBase();
-		configDao.persist(mailBoxSchedProfile);
-
-		// Constructing response
 		AddProfileToMailBoxResponseDTO serviceResponse = new AddProfileToMailBoxResponseDTO();
+		ResponseDTO response = null;
 
-		ResponseDTO response = new ResponseDTO();
-		response.setMessage(MailBoxConstants.ADD_PROFILE_SUCCESS);
-		response.setStatus(MailBoxConstants.SUCCESS);
+		try {
 
-		serviceResponse.setResponse(response);
-		serviceResponse.setMailboxProfileLinkGuid(String.valueOf(mailBoxSchedProfile.getPrimaryKey()));
+			if (mailboxGuid == null || mailboxGuid == "") {
+				throw new MailBoxServicesException("MailBox cannot be found");
+			}
+			ScheduleProfilesRef profile = new ScheduleProfilesRef();
+			// Getting the existing mailbox from given GUID
+			MailBoxConfigurationDAO config = new MailBoxConfigurationDAOBase();
+			MailBox mailbox = config.find(MailBox.class, mailboxGuid);
+			if (null == mailbox) {
+				throw new MailBoxServicesException("The given mailbox GUID is not available in the system.");
+			}
 
-		LOG.info("Exit from add profile to mailbox.");
+			// Construct MailBoxSchedProfile entity
+			MailBoxSchedProfile mailBoxSchedProfile = new MailBoxSchedProfile();
+			mailBoxSchedProfile.setScheduleProfilesRef(profile);
+			mailBoxSchedProfile.setMailbox(mailbox);
+			request.copyToEntity(mailBoxSchedProfile);
+
+			// Persist the entity
+			MailBoxScheduleProfileConfigurationDAO configDao = new MailBoxScheduleProfileConfigurationDAOBase();
+			configDao.persist(mailBoxSchedProfile);
+
+			// Constructing response
+			response = new ResponseDTO();
+			response.setMessage(MailBoxConstants.ADD_PROFILE_SUCCESS);
+			response.setStatus(MailBoxConstants.SUCCESS);
+
+			serviceResponse.setResponse(response);
+			serviceResponse.setMailboxProfileLinkGuid(String.valueOf(mailBoxSchedProfile.getPrimaryKey()));
+
+			LOG.info("Exit from add profile to mailbox.");
+
+		} catch (Exception e) {
+
+			response = new ResponseDTO();
+			response.setMessage(MailBoxConstants.ADD_PROFILE_FAILURE + e.getMessage());
+			response.setStatus(MailBoxConstants.FAILURE);
+
+			serviceResponse.setResponse(response);
+		}
 		return serviceResponse;
 
 	}
@@ -88,23 +105,44 @@ public class ProfileConfigurationService {
 	public DeactivateMailboxProfileLinkResponseDTO deactivateMailboxProfileLink(String mailboxGuid, String linkGuid) {
 
 		LOG.info("Deactivate Profile from mailbox");
-
-		// Getting the existing mailBoxSchedProfile from given link GUID
-		MailBoxScheduleProfileConfigurationDAO configDao = new MailBoxScheduleProfileConfigurationDAOBase();
-		MailBoxSchedProfile mailBoxSchedProfile = configDao.find(MailBoxSchedProfile.class, linkGuid);
-		configDao.remove(mailBoxSchedProfile);
-
-		// Constructing service response
 		DeactivateMailboxProfileLinkResponseDTO serviceResponse = new DeactivateMailboxProfileLinkResponseDTO();
+		ResponseDTO response = null;
 
-		ResponseDTO response = new ResponseDTO();
-		response.setMessage(MailBoxConstants.DEACTIVATE_MAIBOX_PROFILE_SUCCESS);
-		response.setStatus(MailBoxConstants.SUCCESS);
+		try {
 
-		serviceResponse.setResponse(response);
-		serviceResponse.setMailboxProfileLinkGuid(String.valueOf(linkGuid));
+			if (mailboxGuid == null || mailboxGuid == "") {
+				throw new MailBoxServicesException("MailBox cannot be found");
+			} else if (linkGuid == null || mailboxGuid == "") {
+				throw new MailBoxServicesException("MailBox cannot be found");
+			}
+			// Getting the existing mailBoxSchedProfile from given link GUID
+			MailBoxScheduleProfileConfigurationDAO configDao = new MailBoxScheduleProfileConfigurationDAOBase();
+			MailBoxSchedProfile mailBoxSchedProfile = configDao.find(MailBoxSchedProfile.class, linkGuid);
 
-		LOG.info("Exit from deactivate mailbox profile.");
+			if (null == mailBoxSchedProfile) {
+				throw new MailBoxServicesException("No such MailBox Profile found");
+			}
+			configDao.remove(mailBoxSchedProfile);
+
+			// Constructing service response
+
+			response = new ResponseDTO();
+			response.setMessage(MailBoxConstants.DEACTIVATE_MAIBOX_PROFILE_SUCCESS);
+			response.setStatus(MailBoxConstants.SUCCESS);
+
+			serviceResponse.setResponse(response);
+			serviceResponse.setMailboxProfileLinkGuid(String.valueOf(linkGuid));
+
+			LOG.info("Exit from deactivate mailbox profile.");
+
+		} catch (Exception e) {
+
+			response = new ResponseDTO();
+			response.setMessage(MailBoxConstants.ADD_PROFILE_FAILURE + e.getMessage());
+			response.setStatus(MailBoxConstants.FAILURE);
+
+			serviceResponse.setResponse(response);
+		}
 		return serviceResponse;
 	}
 }
